requirements.txt
streamlit==1.49.1
pandas==2.3.2
numpy==2.3.3
pyarrow==21.0.0
altair==5.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1